extracted sessions functions into its own module
