+++ lib/cambiatus/auth/auth.ex
@@ -1,10 +1,8 @@
 defmodule Cambiatus.Auth do
   @moduledoc "Cambiatus Account Authentication and Signup Manager"
 
-  @valid_token 2 * 24 * 3600
   import Ecto.Query
 
-  alias CambiatusWeb.AuthToken
   alias Cambiatus.{
     Accounts,
     Accounts.User,
@@ -15,9 +13,7 @@ defmodule Cambiatus.Auth do
   alias Cambiatus.Auth.{
     Invitation,
     InvitationId,
-    Ecdsa,
-    Phrase,
-    UserToken
+    Session
   }
 
   @contract Application.get_env(:cambiatus, :contract)
@@ -52,170 +48,24 @@ defmodule Cambiatus.Auth do
 
   def gen_auth_phrase(user) do
     user
-    |> create_phrase()
-  end
-
-  def create_phrase(user) do
-    if get_user_token(%{account: user.account, filter: :session}) do
-      {:error, "Session already exists"}
-    else
-      phrase = Phrase.generate()
-
-      %UserToken{}
-      |> UserToken.changeset(%{phrase: phrase, context: "auth"}, user)
-      |> Repo.insert()
-      |> case do
-        {:ok, _} -> {:ok, phrase}
-        {:error, _err} -> {:error, "Phrase exists already"}
-      end
-    end
+    |> Session.create_phrase()
   end
 
   def verify(token) do
     token
-    |> verify_session_token()
-    |> get_account_from_token()
-  end
-
-  def verify_session_token(token) do
-    %{token: token, filter: :session}
-    |> get_user_token()
-    |> case do
-      user_token when is_map(user_token) ->
-        {AuthToken.verify(user_token.token, @valid_token), user_token}
-
-      nil ->
-        {:error, "Session not found"}
-    end
-  end
-
-   def verify_session_token(account, token) do
-    %{account: account, filter: :session}
-    |> get_user_token()
-    |> Map.values()
-    |> Enum.member?(token)
-    |> case do
-      true ->
-        AuthToken.verify(token, @valid_token)
-
-      false ->
-        delete_user_token(%{account: account, filter: :auth})
-        {:error, "Session not found"}
-    end
-  end
-
-  def get_account_from_token({{:ok, _}, user_token}) do
-    Cambiatus.Accounts.get_user(user_token.user_id)
-  end
-
-  def get_account_from_token({{:error, _}, _}) do
-    {:error, "Invalid session"}
-  end
-
-  def get_account_from_token({:error, _}) do
-    {:error, "Invalid session"}
+    |> Session.verify_session_token()
+    |> Session.get_account_from_token()
   end
 
   def verify_signature(phrase, signature) do
     %{phrase: phrase, filter: :auth}
-    |> get_user_token()
-    |> verify_signature_helper(phrase, signature)
+    |> Session.get_user_token()
+    |> Session.verify_signature_helper(phrase, signature)
   end
 
-  defp verify_signature_helper(nil, _phrase, _signature) do
-    {:error, "No phrase found"}
-  end
-  defp verify_signature_helper(%{user_id: account}, phrase, signature) do
-    if Ecdsa.verify_signature(account, signature, phrase) do
-      user = Accounts.get_user(account)
-      token = create_session(user)
-      delete_user_token(%{account: account, filter: :auth})
+  def create_session(user), do: Session.create_session(user)
 
-      {:ok, {user, token}}
-    else
-      {:error, "Invalid signature"}
-    end
-  end
-
-  def create_session({:error, _} = err), do: err
-
-  def create_session(user) do
-    token = AuthToken.gen_token(%{account: user.account})
-
-    %UserToken{}
-    |> UserToken.changeset(%{token: token, context: "session"}, user)
-    |> Repo.insert()
-    |> case do
-      {:ok, _} -> token
-      {:error, _err} -> :error
-    end
-  end
-
-  def get_user_token(%{account: account, filter: :session}) do
-    from("user_tokens",
-      where: [context: "session", user_id: ^account],
-      select: [:context, :user_id, :token]
-    )
-    |> Repo.one()
-  end
-
-  def get_user_token(%{token: token, filter: :session}) do
-    from("user_tokens",
-      where: [context: "session", token: ^token],
-      select: [:context, :user_id, :token]
-    )
-    |> Repo.one()
-  end
-
-  def get_user_token(%{phrase: phrase, filter: :auth}) do
-    from("user_tokens",
-      where: [context: "auth", phrase: ^phrase],
-      select: [:context, :user_id, :token, :phrase]
-    )
-    |> Repo.one()
-  end
-
-  def get_user_token(%{account: account, filter: :auth}) do
-    from("user_tokens",
-      where: [context: "auth", user_id: ^account],
-      select: [:context, :user_id, :token, :phrase]
-    )
-    |> Repo.one()
-  end
-
-  def delete_user_token(%{account: account, filter: :session}) do
-    from("user_tokens",
-      where: [context: "session", user_id: ^account]
-    )
-    |> IO.inspect(label: "RESULT")
-    |> Repo.delete_all()
-  end
-
-  def delete_user_token(%{account: account, filter: :auth}) do
-    from("user_tokens",
-      where: [context: "auth", user_id: ^account]
-    )
-    |> Repo.delete_all()
-  end
-
-  def delete_all_user_token() do
-    [:session, :auth]
-    |> Enum.each(&delete_all_user_token(&1))
-  end
-
-  def delete_all_user_token(:auth) do
-    from("user_tokens",
-      where: [context: "auth"]
-    )
-    |> Repo.delete_all()
-  end
-
-  def delete_all_user_token(:session) do
-    from("user_tokens",
-      where: [context: "session"]
-    )
-    |> Repo.delete_all()
-  end
+  def delete_user_token(args),do: Session.delete_user_token(args)
 
   def netlink({:ok, user}, invitation_id) do
     with {:ok, invitation} <- Auth.get_invitation(invitation_id),
